// src/types.ts
export interface MintsDataReponse {
  tokenMint?: string;
  solMint?: string;
}
// ...
export interface WebSocketRequest {
  jsonrpc: string;
  id: number;
  method: string;
  params: unknown[];
}